requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
pennylane
qiskit
dataset